+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 * feat(PageCard): new component
 * feat(PageGrid): new component
 * feat(PageColumns): new component
+* feat(PageList): new component
 * fix(ProseLi,ProseOl,ProseUl,ProseP): fix color && size
 * fix(ProseH1\H2\H3\H4\H5\H6): add support hash
 * fix(SidebarLayout): improve theme, import Button, add slot 'content-right'